server/systime_mon: move server/systime_mon.go to pkg/systimemon/
Other utilities are almost all there.
